refactor: Changing representation of device addresses
Was represented as 'struct nvm_address' which did not make much sense,
since the functions taking `struct nvm_address` expect that the address
is on generic-format.

+++ cli/addr.c
@@ -160,7 +160,7 @@ int cmd_fmt(NVM_CLI_CMD_ARGS *args, int flags)
 int cmd_to_lba(NVM_CLI_CMD_ARGS *args, int flags)
 {
 	for (int i = 0; i < args->naddrs; ++i) {
-		size_t lba;
+		uint64_t lba;
 
 		lba = nvm_addr_gen2lba(args->dev, args->addrs[i]);
